Dashboard/requirements.txt
nats-py
influxdb-client
aiohttp
